refactor(agentic-ai): extract AzureAuthentication into shared package
Move the AzureAuthentication sealed type and its API-key/client-credentials
records out of AzureOpenAiProviderConfiguration into the shared/ package so a
future provider (Azure AI Foundry) can reference the same auth type without
duplication. JSON subtype IDs and template annotations remain identical;
generated element templates are byte-identical.
