+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/model/request/provider/AzureOpenAiProviderConfiguration.java
@@ -8,13 +8,11 @@ package io.camunda.connector.agenticai.aiagent.model.request.provider;
 
 import static io.camunda.connector.agenticai.aiagent.model.request.provider.AzureOpenAiProviderConfiguration.AZURE_OPENAI_ID;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import io.camunda.connector.agenticai.aiagent.model.request.provider.shared.AzureAuthentication;
 import io.camunda.connector.agenticai.aiagent.model.request.provider.shared.HttpUrl;
 import io.camunda.connector.agenticai.aiagent.model.request.provider.shared.TimeoutConfiguration;
 import io.camunda.connector.api.annotation.FEEL;
 import io.camunda.connector.generator.java.annotation.FeelMode;
-import io.camunda.connector.generator.java.annotation.TemplateDiscriminatorProperty;
 import io.camunda.connector.generator.java.annotation.TemplateProperty;
 import io.camunda.connector.generator.java.annotation.TemplateSubType;
 import jakarta.validation.Valid;
@@ -47,88 +45,6 @@ public record AzureOpenAiProviderConfiguration(@Valid @NotNull AzureOpenAiConnec
       @Valid @NotNull AzureAuthentication authentication,
       @Valid TimeoutConfiguration timeouts,
       @Valid @NotNull AzureOpenAiModel model) {}
-
-  @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-  @JsonSubTypes({
-    @JsonSubTypes.Type(
-        value = AzureAuthentication.AzureApiKeyAuthentication.class,
-        name = "apiKey"),
-    @JsonSubTypes.Type(
-        value = AzureAuthentication.AzureClientCredentialsAuthentication.class,
-        name = "clientCredentials")
-  })
-  @TemplateDiscriminatorProperty(
-      label = "Authentication",
-      group = "provider",
-      name = "type",
-      defaultValue = "apiKey",
-      description = "Specify the Azure OpenAI authentication strategy.")
-  public sealed interface AzureAuthentication {
-    @TemplateSubType(id = "apiKey", label = "API key")
-    record AzureApiKeyAuthentication(
-        @NotBlank
-            @TemplateProperty(
-                group = "provider",
-                label = "API key",
-                type = TemplateProperty.PropertyType.String,
-                feel = FeelMode.optional,
-                constraints = @TemplateProperty.PropertyConstraints(notEmpty = true))
-            String apiKey)
-        implements AzureAuthentication {
-
-      @Override
-      public @NotNull String toString() {
-        return "AzureApiKeyAuthentication{apiKey=[REDACTED]}";
-      }
-    }
-
-    @TemplateSubType(id = "clientCredentials", label = "Client credentials")
-    record AzureClientCredentialsAuthentication(
-        @NotBlank
-            @TemplateProperty(
-                group = "provider",
-                label = "Client ID",
-                description = "ID of a Microsoft Entra application",
-                type = TemplateProperty.PropertyType.String,
-                feel = FeelMode.optional,
-                constraints = @TemplateProperty.PropertyConstraints(notEmpty = true))
-            String clientId,
-        @NotBlank
-            @TemplateProperty(
-                group = "provider",
-                label = "Client secret",
-                description = "Secret of a Microsoft Entra application",
-                type = TemplateProperty.PropertyType.String,
-                feel = FeelMode.optional,
-                constraints = @TemplateProperty.PropertyConstraints(notEmpty = true))
-            String clientSecret,
-        @NotBlank
-            @TemplateProperty(
-                group = "provider",
-                label = "Tenant ID",
-                description =
-                    "ID of a Microsoft Entra tenant. Details in the <a href=\"https://learn.microsoft.com/en-us/entra/fundamentals/how-to-find-tenant\" target=\"_blank\">documentation</a>.",
-                type = TemplateProperty.PropertyType.String,
-                feel = FeelMode.optional)
-            String tenantId,
-        @TemplateProperty(
-                group = "provider",
-                label = "Authority host",
-                description =
-                    "Authority host URL for the Microsoft Entra application. Defaults to <code>https://login.microsoftonline.com</code>. This can also contain an OAuth 2.0 token endpoint.",
-                type = TemplateProperty.PropertyType.String,
-                feel = FeelMode.optional,
-                optional = true)
-            String authorityHost)
-        implements AzureAuthentication {
-
-      @Override
-      public String toString() {
-        return "AzureClientCredentialsAuthentication{clientId=%s, clientSecret=[REDACTED], tenantId=%s, authorityHost=%s}"
-            .formatted(clientId, tenantId, authorityHost);
-      }
-    }
-  }
 
   public record AzureOpenAiModel(
       @NotBlank
